fix(models): normalize Mamba provider stack spec
